Add: Daily commands backup.
